+++ packages/react-api/src/api/stats.js
@@ -1,5 +1,5 @@
 import { checkCredentials, makeCall } from './common';
-import { MAP_TYPES, API_VERSIONS } from '@deck.gl/carto';
+import { MAP_TYPES, API_VERSIONS } from '@deck.gl/carto/typed';
 import { getTileJson } from './tilejson';
 import { InvalidColumnError, _assert as assert } from '@carto/react-core/';