+++ packages/react-api/src/api/tilejson.js
@@ -1,6 +1,6 @@
 import { checkCredentials, CLIENT_ID } from './common';
-import { MAP_TYPES, API_VERSIONS, fetchLayerData, FORMATS } from '@deck.gl/carto';
-import { _assert as assert } from '@carto/react-core/';
+import { MAP_TYPES, API_VERSIONS, fetchLayerData, FORMATS } from '@deck.gl/carto/typed';
+import { _assert as assert } from '@carto/react-core';
 
 /**
  * Get the TileJson for static tilesets